# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  portfolio-website-backend_internal_dto.ProjectData:
    properties:
      additional_data:
        additionalProperties: true
        type: object
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      project_category_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  portfolio-website-backend_internal_dto.PublicDataResponse:
    properties:
      about:
        additionalProperties: true
        type: object
      availability:
        type: string
      avatar:
        type: string
      email:
        type: string
      featuredSkills:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_dto.SkillData'
        type: array
      heroStats:
        additionalProperties: true
        type: object
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      projectCategories:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_models.ProjectCategory'
        type: array
      projects:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_dto.ProjectData'
        type: array
      projectsSection:
        additionalProperties: true
        type: object
      reviews:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_models.Review'
        type: array
      skillGroups:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_dto.SkillGroupData'
        type: array
      skillsSection:
        additionalProperties: true
        type: object
      socialLinks:
        additionalProperties: true
        type: object
      surname:
        type: string
      timelineData:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_dto.TimelineItem'
        type: array
      timelineSection:
        additionalProperties: true
        type: object
      title:
        type: string
    type: object
  portfolio-website-backend_internal_dto.SkillData:
    properties:
      color:
        type: string
      icon:
        type: string
      id:
        type: string
      name:
        type: string
      proficiency:
        type: integer
    type: object
  portfolio-website-backend_internal_dto.SkillGroupData:
    properties:
      name:
        type: string
      skills:
        items:
          $ref: '#/definitions/portfolio-website-backend_internal_dto.SkillData'
        type: array
    type: object
  portfolio-website-backend_internal_dto.TimelineItem:
    properties:
      company:
        type: string
      degree:
        type: string
      description:
        type: string
      id:
        type: string
      institution:
        type: string
      period:
        type: string
      title:
        type: string
      type:
        type: string
      year:
        type: integer
    type: object
...
  portfolio-website-backend_internal_models.JSONMap:
    additionalProperties: true
    type: object
...
  portfolio-website-backend_internal_models.ProjectCategory:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_visible:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  portfolio-website-backend_internal_models.Review:
    properties:
      avatar:
        type: string
      company:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_visible:
        type: boolean
      name:
        type: string
      position:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
    type: object
  portfolio-website-backend_internal_models.User:
    properties:
      about:
        $ref: '#/definitions/portfolio-website-backend_internal_models.JSONMap'
      availability:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      featured_skill_ids:
        items:
          type: string
        type: array
      id:
        type: string
      location:
        type: string
      name:
        description: Optional fields
        type: string
      phone:
        type: string
      social_links:
        $ref: '#/definitions/portfolio-website-backend_internal_models.JSONMap'
      surname:
        type: string
      title:
        type: string
      updated_at:
        type: string
      username:
        description: Required fields
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is the backend API for my portfolio website.
  title: Portfolio Website API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login and get an access token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - users
  /profile:
    get:
      consumes:
      - application/json
      description: Get the profile of the currently logged in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the profile of the currently logged in user
      parameters:
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
...
  /public-data/{user_id}:
    get:
      consumes:
      - application/json
      description: Get public portfolio data for a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolio-website-backend_internal_dto.PublicDataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get public data
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"